requirements.txt
python-telegram-bot==20.3
wakeonlan==2.1.0